tests/wpt/tests/css/css-page/parsing/page-orientation-computed.tentative.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-page-3/#page-orientation-prop">
<div id="elm" style="page-orientation:rotate-right;"></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(()=> {
      assert_equals(getComputedStyle(elm).pageOrientation, "");
  }, "page-orientation is not a property (only a descriptor)");
</script>
